Update styles, fix php script bug

+++ api/shoot.php
@@ -61,7 +61,7 @@
 
 	$hit = $area->point_within($x, $y);
 
-	$end_time = microtime(true);
+	$end_time = (int) microtime(true);
 	$script_time = number_format($end_time - $start_time, 5, '.', '');
 	$end_datetime = date("d/m/Y H:i:s", $end_time);
 ?>
